dav/dav2/website-db-argocd-kustomize-kyverno-grafana-loki-tempo/app/requirements.txt
fastapi
uvicorn
jinja2
psycopg2-binary
prometheus-fastapi-instrumentator
